Graphs and unfinished proof of taken levels
I finished two graphs. The first one is the graph of enrolled students
and the second one is of registered groups of all registered cycles.
Graph of registered groups still needs to be checked, and we probably
need to eliminate the proof of taken levels to not have any problems.